+++ REFACTOR_NOTE.md
@@ -1,0 +1,5 @@
+Internal:
+- Renamed `TooManyArguments` Exception to `TooManyArgumentsError` for name consistency
+- Renamed `SubCommandNotFound` Exception to `SubCommandNotFoundError` for name consistency
+
+Deprecation:
